Make environment.env optional in user data

+++ deploy/packer/scripts/chroot.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/bash
 # 
 # VAZ Projects
 # 
@@ -164,26 +164,28 @@ EOF
 # Init scripts.
 #-------------------------------------------------------------------------------
 cat > /usr/local/lib/instanceScriptsSetup.sh << 'EOF'
-#!/bin/bash
+#!/usr/bin/bash
 # Grow root partition.
 disk="/dev/$(lsblk -nro MOUNTPOINT,PKNAME | grep '^/ ' | cut -d ' ' -f2)"
 partition="/dev/$(lsblk -nro MOUNTPOINT,KNAME | grep '^/ ' | cut -d ' ' -f2)"
 
+# Resize root partition.
 sgdisk --move-second-header ${disk}
 sgdisk --delete 2 ${disk}
 sgdisk --new 2:0:0 --change-name 2:'Root' ${disk}
 
+# Reload partitions.
 partx -u ${disk}
 
+# Resize filesystem.
 resize2fs ${partition}
 
 
 # Download user data.
 mkdir /tmp/deploy && cd ${_}
 curl -s http://169.254.169.254/latest/user-data | tar -xz
-# TODO: Make Optional.
 mv per*.sh /usr/local/lib/
-mv environment.env /etc/environment
+test -f environment.env && mv environment.env /etc/environment || true
 EOF
 #-------------------------------------------------------------------------------
 chmod +x /usr/local/lib/instanceScriptsSetup.sh
